+++ hosts/darwin/palantir/default.nix
@@ -1,0 +1,18 @@
+{ users, ... }:
+{
+  imports = [
+    ../../common/darwin
+  ];
+
+  users = [
+    users.noghartt
+  ];
+
+  nixpkgs.hostPlatform = "aarch64-darwin";
+  device.type = "laptop";
+
+  system = {
+    primaryUser = "noghartt";
+    stateVersion = 7;
+  };
+}